n1sdp: add config files for DMC620 & DDR PHY and include in build
